New src/widget/view/shared module
