+++ src/widget/view/mod.rs
@@ -8,18 +8,14 @@
 //! View widgets exist as a view over some shared data.
 
 use super::Label;
-#[allow(unused)]
-use kas::event::Manager;
-use kas::event::UpdateHandle;
 use kas::prelude::*;
-use std::cell::RefCell;
-use std::fmt::Debug;
-use std::rc::Rc;
 
 // mod list;
+mod shared;
 mod single;
 
 // pub use list::ListView;
+pub use shared::{Accessor, AccessorShared, SharedConst, SharedRc};
 pub use single::SingleView;
 
 /// View widgets
@@ -84,117 +80,4 @@ impl DefaultView for String {
 }
 impl<'a> DefaultView for &'a str {
     type Widget = Label<String>;
-}
-
-/// Base trait required by view widgets
-// Note: we require Debug + 'static to allow widgets using this to implement
-// WidgetCore, which requires Debug + Any.
-pub trait Accessor<I>: Debug + 'static {
-    type Item;
-
-    /// Size descriptor
-    ///
-    /// Note: for `I == ()` we consider `()` a valid index; in other cases we
-    /// usually expect `index < accessor.len()` (for each component).
-    fn len(&self) -> I;
-
-    /// Access data by index
-    fn get(&self, index: I) -> Self::Item;
-
-    /// Get an update handle, if any is used
-    ///
-    /// Widgets may use this `handle` to call `mgr.update_on_handle(handle, self.id())`.
-    fn update_handle(&self) -> Option<UpdateHandle> {
-        None
-    }
-}
-
-/// Extension trait for shared data for view widgets
-pub trait AccessorShared<I>: Accessor<I> {
-    /// Set data at the given index
-    ///
-    /// The caller should call [`Manager::trigger_update`] using the returned
-    /// update handle, using an appropriate transformation of the index for the
-    /// payload (the transformation defined by implementing view widgets).
-    /// Calling `trigger_update` is unnecessary before the UI has been started.
-    fn set(&self, index: I, value: Self::Item) -> UpdateHandle;
-}
-
-/// Wrapper for shared constant data
-///
-/// This may be useful with static data, e.g. `[&'static str]`.
-#[derive(Clone, Debug, Default)]
-pub struct SharedConst<T: Debug + 'static + ?Sized>(T);
-
-impl<T: Clone + Debug + 'static + ?Sized> SharedConst<T> {
-    /// Construct with given data
-    pub fn new(data: T) -> Self {
-        SharedConst(data)
-    }
-}
-
-impl<T: Clone + Debug + 'static> Accessor<()> for SharedConst<T> {
-    type Item = T;
-    fn len(&self) -> () {
-        ()
-    }
-    fn get(&self, _: ()) -> T {
-        self.0.clone()
-    }
-}
-
-impl<T: Clone + Debug + 'static + ?Sized> Accessor<usize> for SharedConst<[T]> {
-    type Item = T;
-    fn len(&self) -> usize {
-        self.0.len()
-    }
-    fn get(&self, index: usize) -> T {
-        self.0[index].to_owned()
-    }
-}
-
-/// Wrapper for single-thread shared data
-#[derive(Clone, Debug)]
-pub struct SharedRc<T: Clone + Debug + 'static> {
-    handle: UpdateHandle,
-    data: Rc<RefCell<T>>,
-}
-
-impl<T: Default + Clone + Debug + 'static> Default for SharedRc<T> {
-    fn default() -> Self {
-        SharedRc {
-            handle: UpdateHandle::new(),
-            data: Default::default(),
-        }
-    }
-}
-
-impl<T: Clone + Debug + 'static> SharedRc<T> {
-    /// Construct with given data
-    pub fn new(data: T) -> Self {
-        SharedRc {
-            handle: UpdateHandle::new(),
-            data: Rc::new(RefCell::new(data)),
-        }
-    }
-}
-
-impl<T: Clone + Debug + 'static> Accessor<()> for SharedRc<T> {
-    type Item = T;
-    fn len(&self) -> () {
-        ()
-    }
-    fn get(&self, _: ()) -> T {
-        self.data.borrow().to_owned()
-    }
-    fn update_handle(&self) -> Option<UpdateHandle> {
-        Some(self.handle)
-    }
-}
-
-impl<T: Clone + Debug + 'static> AccessorShared<()> for SharedRc<T> {
-    fn set(&self, _: (), value: T) -> UpdateHandle {
-        *self.data.borrow_mut() = value;
-        self.handle
-    }
 }
